refactor: remove legacy code
